+++ app/src/main/java/com/limelight/nvstream/NvConnectionListener.java
@@ -12,5 +12,5 @@ public interface NvConnectionListener {
     void displayMessage(String message);
     void displayTransientMessage(String message);
 
-    void rumble(short controllerNumber, short lowFreqMotor, short highFreqMotor);
+    void rumble(int controllerNumber, int lowFreqMotor, int highFreqMotor);
 }
